+++ shared/contants/index.ts
@@ -1,6 +1,6 @@
 export const NAVIGATION_LIST = [
   { label: '강화하기', url: '/' },
-  { label: '행운의 시바', url: '/luckyshiba' },
+  { label: '운좋은 시바', url: '/luckyshiba' },
 ];
 
 export const COLLECTIONS = {
